requirements.txt
certifi==2020.12.5
chardet==4.0.0
django-simple-rest==1.4.1
idna==2.10
mimeparse==0.1.3
requests==2.25.1
urllib3==1.26.2
wheel==0.36.2
